2004-07-17 Erik Hjortsberg <erik@hysteriskt.nu> usr/components/ogre: *avatar now snaps clientside to the terrain, we need however to increase the resolution of the terrain; perhaps we should wait until ogre 0.15 (because of both TerrainSceneManager and PagingTerrainManager being updated in the next release)

+++ src/components/ogre/terrain/TerrainGenerator.cpp
@@ -60,12 +60,12 @@ void TerrainGenerator::prepareSegments(long segmentXStart, long segmentZStart, l
 }
 
 
-float TerrainGenerator::getHeight(long ogreX, long ogreZ)
+float TerrainGenerator::getHeight(float ogreX, float ogreZ)
 {
 	
 	//convert our ogre coordinates to atlas
-	long atlasX = ogreX;
-	long atlasY = -ogreZ;
+	float atlasX = ogreX;
+	float atlasY = -ogreZ;
 	
 	float height = mTerrain.get(atlasX, atlasY);
 	return height;
